+++ src/Google/GoogleEvent.php
@@ -6,15 +6,12 @@ namespace Nexendrie\EventCalendar\Google;
 
 /**
  * Represents single event from Google Calendar
- *
- * @property-read string $id
- * @property-read string $creator
  */
 final class GoogleEvent
 {
     use \Nette\SmartObject;
 
-    private string $id;
+    public string $id;
     public string $status;
     public string $htmlLink;
     public \DateTime $created;
@@ -22,22 +19,12 @@ final class GoogleEvent
     public string $summary;
     public ?string $location = null;
     public ?string $description = null;
-    private string $creator = '';
+    public string $creator = '';
     public \DateTime $start;
     public \DateTime $end;
     
     public function __construct(string $id)
     {
         $this->id = $id;
-    }
-    
-    public function getId(): string
-    {
-        return $this->id;
-    }
-
-    public function getCreator(): string
-    {
-        return $this->creator;
     }
 }